+++ js/index.js
@@ -1,26 +1,5 @@
-function makeCanvas() {
-  const canvas = document.createElement("div");
-  canvas.classList.add("canvas");
-  // canvas.classList.add("move");
-  document.querySelector("body").appendChild(canvas);
-  for (let i = 0; i < 2000; i++) {
-    const rows = document.createElement("div");
-    rows.classList.add("rows");
-    rows.innerHTML = `
-      <div data-index='${i + "0"}' class="white-tiles"></div>
-      <div data-index='${i + "1"}' class="white-tiles"></div>
-      <div data-index='${i + "2"}' class="white-tiles"></div>
-      <div data-index='${i + "3"}' class="white-tiles"></div>
-    `;
-    const randPosition = Math.floor(Math.random() * 4);
-    rows
-      .querySelectorAll(".white-tiles")
-      [randPosition].classList.add("black-tile");
-    canvas.appendChild(rows);
-    canvas.scrollTop = canvas.scrollHeight;
-  }
-}
+import { makeCanvas } from "./canvas.js";
 
-makeCanvas();
+const canvas = document.createElement("div");
 
-function detectClickOnTiles() {}
+makeCanvas(canvas);
